Mesa: Use __attribute__((destructor)), not atexit(3)
On Linux, atexit(3) registered functions are called at program exit or
during module unload. The latter is a Glibc extension not supported by
FreeBSD. This means that, on FreeBSD, the registered function could be
called after the module was unloaded, causing the application to crash.

The patch replaces the use of atexit(3) by the compiler attribute
`destructor`.

This fixes a segfault with clinfo(1) and darktable.